requirements.txt
voicevox-client 
requests 
sounddevice 
numpy 
aiohttp
pydub